add link checker to prod, fix broken links
This doesn't block anything in CI yet, it just runs
after production is deployed. It also completely ignores anything
on community, although that's something to tackle as well at some point.

Here's an example of a build failing due to a few broken links in prod
(which are also incidentally fixed in this PR)

https://circleci.com/gh/dexhorthy/help-center/7